fix: update product proto to add orderId

+++ product/pkg/models/product.go
@@ -1,10 +1,10 @@
 package models
 
 type Product struct {
-	Id                int64            `json:"id" gorm:"primaryKey"`
+	Id                int32            `json:"id" gorm:"primaryKey"`
 	Name              string           `json:"name"`
-	Stock             int64            `json:"stock"`
-	Price             int64            `json:"price"`
+	Stock             int32            `json:"stock"`
+	Price             float64          `json:"price"`
 	Sku               string           `json:"sku"`
 	StockDecreaseLogs StockDecreaseLog `gorm:"foreignKey:ProductId"`
 }
